+++ src/main/java/com/mosaicchurchaustin/oms/controllers/CustomerController.java
@@ -7,6 +7,8 @@ import com.mosaicchurchaustin.oms.services.CustomerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -54,7 +56,16 @@ public class CustomerController {
 
     @ResponseBody
     @GetMapping(path = "/customer", produces = MediaType.APPLICATION_JSON_VALUE)
-    public Page<CustomerResponse> getCustomersPaged(final Pageable pageable) {
-        return customerService.getCustomers(pageable).map(CustomerResponse::from);
+    public Page<CustomerResponse> getCustomersPaged(
+            @PageableDefault(sort = { "lastName", "firstName" }, direction = Sort.Direction.ASC)
+            final Pageable pageable,
+            @RequestParam(value = "name", required = false)
+            final String name,
+            @RequestParam(value = "flagged", required = false)
+            final Boolean flagged
+    ) {
+        return customerService
+                .getCustomers(pageable, name, flagged)
+                .map(CustomerResponse::from);
     }
 }