+++ src/main/java/com/mosaicchurchaustin/oms/repositories/CustomerRepository.java
@@ -3,6 +3,7 @@ package com.mosaicchurchaustin.oms.repositories;
 import com.mosaicchurchaustin.oms.data.entity.customer.CustomerEntity;
 import com.mosaicchurchaustin.oms.data.projections.CustomerSearchProjection;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
@@ -11,7 +12,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<CustomerEntity, Long> {
+public interface CustomerRepository extends JpaRepository<CustomerEntity, Long>, JpaSpecificationExecutor<CustomerEntity> {
 
     Optional<CustomerEntity> findByFirstNameAndLastName(String firstName, String lastName);
     Optional<CustomerEntity> findByUuid(String uuid);